+++ src/dbtarzan/db/actor/CopyWorker.scala
@@ -9,7 +9,7 @@ import dbtarzan.config.connections.ConnectionData
 import dbtarzan.config.password.EncryptionKey
 import dbtarzan.db.util.ResourceManagement.using
 import dbtarzan.db._
-import dbtarzan.db.foreignkeys.{ ForeignKeyLoader, ForeignKeysToFile }
+import dbtarzan.db.foreignkeys.{ ForeignKeyLoader, ForeignKeysFiles }
 import dbtarzan.messages._
 import dbtarzan.localization.Localization
 
@@ -24,6 +24,7 @@ class CopyWorker(data : ConnectionData, encryptionKey: EncryptionKey, guiActor :
 	def databaseName = data.name
 	val foreignKeyLoader =  new ForeignKeyLoader(connection, DBDefinition(data.schema, data.catalog), localization)
 	val queryLoader = new QueryLoader(connection)
+	val foreignKeysFile = ForeignKeysFiles.forCache(databaseName)
 	/* gets all the tables in the database/schema from the database metadata */
 	private def tableNames() : List[String] = {
 		val meta = connection.getMetaData()
@@ -42,16 +43,15 @@ class CopyWorker(data : ConnectionData, encryptionKey: EncryptionKey, guiActor :
 		val keysForTables = names.map(name => 
 			ForeignKeysForTable(name, foreignKeyLoader.foreignKeys(name))
 			)
-		ForeignKeysToFile.toFile(data.name, ForeignKeysForTableList(keysForTables))
+		foreignKeysFile.toFile(ForeignKeysForTableList(keysForTables))
 	}
 
 
   	def receive = {
 	    case CopyToFile => {
-	    	val fileName = ForeignKeysToFile.fileName(data.name)
-	    	log.info(localization.writingFile(fileName))
+	    	log.info(localization.writingFile(foreignKeysFile.fileName))
 	    	loadAllKeysAndWriteThemToFile()
-	    	log.info(localization.fileWritten(fileName))
+	    	log.info(localization.fileWritten(foreignKeysFile.fileName))
 	    }
   	}
 }
