--- src/dbtarzan/db/foreignkeys/ForeignKeysToFile.scala
@@ -1,41 +1,0 @@
-package dbtarzan.db.foreignkeys
-
-import spray.json._
-import dbtarzan.db.util.FileReadWrite
-import dbtarzan.db.{ ForeignKeyDirection, DBEnumsText, FieldsOnTable, ForeignKey, ForeignKeys, ForeignKeysForTable, ForeignKeysForTableList }
-import java.nio.file.{Path, Paths}
-
-  class ForeignKeyDirectionFormat extends RootJsonFormat[ForeignKeyDirection] {
-    def write(direction: ForeignKeyDirection): JsValue = JsString(DBEnumsText.foreignKeyDirectionToText(direction))
-
-    def read(json: JsValue): ForeignKeyDirection = json match {
-      case JsString("STRAIGHT") => ForeignKeyDirection.STRAIGHT
-	  case JsString("TURNED") => ForeignKeyDirection.TURNED
-      case _ => throw new DeserializationException("ForeignKeyDirection string expected")
-    }
-  }
-
-object ForeignKeysForTableJsonProtocol extends DefaultJsonProtocol {
-  implicit val foreignKeyDirectionFormat = new ForeignKeyDirectionFormat()  
-  implicit val fieldsOnTableFormat = jsonFormat(FieldsOnTable, "table", "fields" )	
-  implicit val foreignKeyFormat = jsonFormat(ForeignKey, "name", "from", "to", "direction")
-  implicit val foreignKeysFormat = jsonFormat(ForeignKeys, "keys")
-  implicit val foreignKeysForTableFormat = jsonFormat(ForeignKeysForTable, "name", "keys")
-  implicit val foreignKeysForTableListFormat = jsonFormat(ForeignKeysForTableList, "keys")
-}
-
-object ForeignKeysToFile {
-	import ForeignKeysForTableJsonProtocol._
-	def fileName(databaseName : String) : Path = Paths.get(databaseName+".fgk")
-
-	def toFile(databaseName : String, list : ForeignKeysForTableList) : Unit =  
-		FileReadWrite.writeFile(fileName(databaseName), list.toJson.prettyPrint)
-	
-	def fromFile(databaseName : String) : ForeignKeysForTableList = {
-		val text = FileReadWrite.readFile(fileName(databaseName))
-		text.parseJson.convertTo[ForeignKeysForTableList]
-	}
-
-	def fileExist(databaseName : String) : Boolean = 
-		FileReadWrite.fileExist(fileName(databaseName))
-}
